editing libft

+++ libft/src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: hebatist <hebatist@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/29 15:39:26 by hebatist          #+#    #+#             */
-/*   Updated: 2024/09/29 15:39:29 by hebatist         ###   ########.fr       */
+/*   Updated: 2024/10/22 21:06:50 by hebatist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	size_t	size;
+	int	size;
 
 	size = 0;
 	while (s[size] != '\0')
